requirements.txt
aiohttp==3.7.4.post0
asn1crypto==1.4.0
async-timeout==3.0.1
attrs==21.2.0
base58==2.1.0
bit==0.7.2
bitarray==1.2.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
coincurve==15.0.1
cytoolz==0.11.0
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
hexbytes==0.2.2
idna==3.3
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==5.2.0
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.19.0
pycparser==2.20
pycryptodome==3.11.0
pyrsistent==0.18.0
python-dotenv==0.19.1
pywin32==302
requests==2.26.0
rlp==2.0.1
six==1.16.0
toolz==0.11.1
typing-extensions==3.10.0.2
urllib3==1.26.7
varint==1.0.2
web3==5.24.0
websockets==9.1
yarl==1.7.0
